MockResponseSender: move it into the test_utils module
Also, hide the internals of test_utils by reexporting the types.
