+++ src/test_utils/mod.rs
@@ -1,1 +1,5 @@
-pub mod correlator;
+pub use self::correlator::MessageEventHandler;
+pub use self::responder::MockResponseSender;
+
+mod correlator;
+mod responder;
