+++ tests/correlator/mod.rs
@@ -1,7 +1,7 @@
 use correlation::CorrelatorFactory;
 use correlation::correlator::Error;
 use correlation::message::MessageBuilder;
-use correlation::test_utils::correlator::MessageEventHandler;
+use correlation::test_utils::MessageEventHandler;
 
 use env_logger;
